+++ packages/class-core/src/LionCore_builtins.g.ts
@@ -27,74 +27,55 @@
  */
 
 
-import {
-    Classifier,
-    Concept,
-    EnumerationLiteral,
-    Interface,
-    Language,
-    PrimitiveType,
-    Property
-} from "@lionweb/core";
+import * as $lwClassCore from "./index.js";
+import * as $lwCore from "@lionweb/core";
+import * as $lwJson from "@lionweb/json";
 
-import {
-    LionWebId
-} from "@lionweb/json";
+export class LionCore_builtinsBase implements $lwClassCore.ILanguageBase {
 
-import {
-    DeltaReceiver,
-    ILanguageBase,
-    INodeBase,
-    NodeBase,
-    NodeBaseFactory
-} from "./index.js";
-
-
-export class LionCore_builtinsBase implements ILanguageBase {
-
-    private readonly _language: Language = new Language("LionCore_builtins", "2023.1", "LionCore-builtins", "LionCore-builtins");
-    get language(): Language {
+    private readonly _language: $lwCore.Language = new $lwCore.Language("LionCore_builtins", "2023.1", "LionCore-builtins", "LionCore-builtins");
+    get language(): $lwCore.Language {
         this.ensureWiredUp();
         return this._language;
     }
 
-    public readonly _String = new PrimitiveType(this._language, "String", "LionCore-builtins-String", "LionCore-builtins-String");
-    get String(): PrimitiveType {
+    public readonly _String = new $lwCore.PrimitiveType(this._language, "String", "LionCore-builtins-String", "LionCore-builtins-String");
+    get String(): $lwCore.PrimitiveType {
         this.ensureWiredUp();
         return this._String;
     }
 
-    public readonly _Boolean = new PrimitiveType(this._language, "Boolean", "LionCore-builtins-Boolean", "LionCore-builtins-Boolean");
-    get Boolean(): PrimitiveType {
+    public readonly _Boolean = new $lwCore.PrimitiveType(this._language, "Boolean", "LionCore-builtins-Boolean", "LionCore-builtins-Boolean");
+    get Boolean(): $lwCore.PrimitiveType {
         this.ensureWiredUp();
         return this._Boolean;
     }
 
-    public readonly _Integer = new PrimitiveType(this._language, "Integer", "LionCore-builtins-Integer", "LionCore-builtins-Integer");
-    get Integer(): PrimitiveType {
+    public readonly _Integer = new $lwCore.PrimitiveType(this._language, "Integer", "LionCore-builtins-Integer", "LionCore-builtins-Integer");
+    get Integer(): $lwCore.PrimitiveType {
         this.ensureWiredUp();
         return this._Integer;
     }
 
-    public readonly _JSON = new PrimitiveType(this._language, "JSON", "LionCore-builtins-JSON", "LionCore-builtins-JSON");
-    get JSON(): PrimitiveType {
+    public readonly _JSON = new $lwCore.PrimitiveType(this._language, "JSON", "LionCore-builtins-JSON", "LionCore-builtins-JSON");
+    get JSON(): $lwCore.PrimitiveType {
         this.ensureWiredUp();
         return this._JSON;
     }
 
-    public readonly _Node = new Concept(this._language, "Node", "LionCore-builtins-Node", "LionCore-builtins-Node", true);
-    get Node(): Concept {
+    public readonly _Node = new $lwCore.Concept(this._language, "Node", "LionCore-builtins-Node", "LionCore-builtins-Node", true);
+    get Node(): $lwCore.Concept {
         this.ensureWiredUp();
         return this._Node;
     }
 
-    public readonly _INamed = new Interface(this._language, "INamed", "LionCore-builtins-INamed", "LionCore-builtins-INamed");
-    get INamed(): Interface {
+    public readonly _INamed = new $lwCore.Interface(this._language, "INamed", "LionCore-builtins-INamed", "LionCore-builtins-INamed");
+    get INamed(): $lwCore.Interface {
         this.ensureWiredUp();
         return this._INamed;
     }
-    private readonly _INamed_name = new Property(this._INamed, "name", "LionCore-builtins-INamed-name", "LionCore-builtins-INamed-name");
-    get INamed_name(): Property {
+    private readonly _INamed_name = new $lwCore.Property(this._INamed, "name", "LionCore-builtins-INamed-name", "LionCore-builtins-INamed-name");
+    get INamed_name(): $lwCore.Property {
         this.ensureWiredUp();
         return this._INamed_name;
     }
@@ -110,14 +91,14 @@ export class LionCore_builtinsBase implements ILanguageBase {
         this._wiredUp = true;
     }
 
-    factory(_receiveDelta?: DeltaReceiver): NodeBaseFactory {
-        return (classifier: Classifier, _id: LionWebId) => {
+    factory(_receiveDelta?: $lwClassCore.DeltaReceiver): $lwClassCore.NodeBaseFactory {
+        return (classifier: $lwCore.Classifier, _id: $lwJson.LionWebId) => {
             const {language} = classifier;
             throw new Error(`can't instantiate ${classifier.name} (key=${classifier.key}): classifier is not known in language ${language.name} (key=${language.key}, version=${language.version})`);
         }
     }
 
-    enumLiteralFrom<EnumType>(enumerationLiteral: EnumerationLiteral): EnumType {
+    enumLiteralFrom<EnumType>(enumerationLiteral: $lwCore.EnumerationLiteral): EnumType {
         const {enumeration} = enumerationLiteral;
         const {language} = enumeration;
         throw new Error(`enumeration with key ${enumeration.key} is not known in language ${language.name} (key=${language.key}, version=${language.version})`);
@@ -135,10 +116,10 @@ export type Integer = number;
 
 export type JSON = unknown;
 
-export abstract class Node extends NodeBase {
+export abstract class Node extends $lwClassCore.NodeBase {
 }
 
-export interface INamed extends INodeBase {
+export interface INamed extends $lwClassCore.INodeBase {
     name: string;
 }