+++ packages/class-core-test-language/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.9.0
+
+* (Regenerate language implementation after changes to `class-core-generator`.)
+
+
 ## 0.8.0
 
 * Regenerate with latest test language definition.