+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.9.0
+
+* Expose types associated with the `measure` function — m.n. `Metrics`.
+* Add a `zeroMetrics` constant, and `mergeMetrics` and `aggregateMetrics` functions, to aggregate `Metrics`.
+* Add a `tryReadSerializationChunk` that plays better with async “versus” throwing exceptions.
+
+
 ## 0.8.0
 
 * Propagate `reference` field of `LionWebJsonReferenceTarget` type now being `null`able.